+++ azalea-client/src/plugins/movement.rs
@@ -65,8 +65,8 @@ impl Plugin for MovementPlugin {
                         .in_set(PhysicsSet)
                         .before(ai_step)
                         .before(azalea_physics::fluids::update_in_water_state_and_do_fluid_pushing),
-                    send_sprinting_if_needed.after(azalea_entity::update_in_loaded_chunk),
                     send_player_input_packet,
+                    send_sprinting_if_needed.after(azalea_entity::update_in_loaded_chunk),
                     send_position.after(PhysicsSet),
                 )
                     .chain(),